+++ onomaturgy/__init__.py
@@ -1,4 +1,5 @@
 from onomaturgy.generators.personal_name import PersonalNameGenerator
 from onomaturgy.generators.place_name import PlaceNameGenerator
+from onomaturgy.generators.company_name import CompanyNameGenerator
 from onomaturgy.generators.simple_name import SimpleNameGenerator
-from onomaturgy.generators.tribal_name import TribalNameGenerator 
+from onomaturgy.generators.tribal_name import TribalNameGenerator